Create default.html
add the default.html layout.  This has a header and a footer.  The head contains a reference to the css stylesheet. The head's title <div>, rather than having a static name, has a liquid tag {{ page.title }}.  The body has href tags to navigate around the site. The container <div> within the body  has a liquid tag for {{ content }}. The footer <div> (nested within the body, has hrefs to email, etc.