Update index.html
change the index quite radically as have moved all the formatting stuff to default.html in the newly created _layouts folder.  front matter takes the place of the html doc, title, etc;  all the hrefs that allowed navigation, reference to a style sheet, footer, were all put in the default.html. Nothing is left here but the blurb <div> (not event the container <div> that used to wrap it)